requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-rds
aws-cdk.aws-secretsmanager
aws-cdk.aws-iam
aws-cdk.aws-autoscaling
aws-cdk.aws_glue